cleanup types & refactor
