+++ core/transactor/tracker/subscription.go
@@ -5,21 +5,7 @@ import (
 	"time"
 
 	"github.com/berachain/offchain-sdk/log"
-
-	coretypes "github.com/ethereum/go-ethereum/core/types"
 )
-
-// Subscriber is an interface that defines methods for handling transaction events.
-type Subscriber interface {
-	// OnSuccess is called when a transaction has been successfully included in a block.
-	OnSuccess(*InFlightTx, *coretypes.Receipt) error
-	// OnRevert is called when a transaction has been reverted.
-	OnRevert(*InFlightTx, *coretypes.Receipt) error
-	// OnStale is called when a transaction becomes stale.
-	OnStale(context.Context, *InFlightTx) error
-	// OnError is called when there is an error with the transaction.
-	OnError(context.Context, *InFlightTx, error)
-}
 
 // Subscription manages a Subscriber, allowing it to be started and stopped.
 type Subscription struct {
